refactor(test): Refactor the tests to include partial runs
